+++ example/lib/src/priorized_event.dart
@@ -1,0 +1,3 @@
+class PriorityEvent {
+  String data = '';
+}