+++ lib/src/builder/dto/extracted_event.dart
@@ -8,10 +8,14 @@ class ExtractedSubscriber {
   /// The handler type
   final SymbolReference subscriber;
 
+  /// The priority of this subscriber
+  final int priority;
+
   /// ExtractedSubscriber constructor
   const ExtractedSubscriber({
     required this.event,
     required this.subscriber,
+    required this.priority,
   });
 
   /// Creates a serializable representation of the subscriber.
@@ -19,6 +23,7 @@ class ExtractedSubscriber {
     return {
       'event': event.toJson(),
       'subscriber': subscriber.toJson(),
+      'priority': priority,
     };
   }
 
@@ -28,6 +33,7 @@ class ExtractedSubscriber {
       event: SymbolReference.fromJson(json['event'] as Map<String, dynamic>),
       subscriber:
           SymbolReference.fromJson(json['subscriber'] as Map<String, dynamic>),
+      priority: json['priority'] as int,
     );
   }
 }